+++ packages/fluentui/docs/src/prototypes/table/InteractiveTable.tsx
@@ -21,7 +21,7 @@ const InteractiveTable = () => {
 
   const contextMenuItems = ['Add to selection', 'Remove', 'Download'];
 
-  const getRowWithContextMenu = tableRow => <MenuButton contextMenu trigger={tableRow} menu={contextMenuItems} />;
+  const withContextMenu = tableRow => <MenuButton contextMenu trigger={tableRow} menu={contextMenuItems} />;
 
   return (
     <>
@@ -53,7 +53,7 @@ const InteractiveTable = () => {
           <Table.Cell content="Picture" key="pic" accessibility={gridHeaderCellBehavior} />
           <Table.Cell content="Age" key="age" accessibility={gridHeaderCellBehavior} />
         </Table.Row>
-        {getRowWithContextMenu(
+        {withContextMenu(
           <Table.Row key="1" accessibility={gridRowBehavior}>
             <Table.Cell content="1" key="1-1" accessibility={gridCellBehavior} />
             <Table.Cell content="Roman van von der Longername" key="1-2" accessibility={gridCellBehavior} />
@@ -61,7 +61,7 @@ const InteractiveTable = () => {
             <Table.Cell content="30 years" key="1-4" accessibility={gridCellBehavior} />
           </Table.Row>
         )}
-        {getRowWithContextMenu(
+        {withContextMenu(
           <Table.Row key="2" accessibility={gridRowBehavior}>
             <Table.Cell content="2" key="1-1" accessibility={gridCellBehavior} />
             <Table.Cell content="Alex" key="1-2" accessibility={gridCellBehavior} />
@@ -69,7 +69,7 @@ const InteractiveTable = () => {
             <Table.Cell content="1 year" key="1-4" accessibility={gridCellBehavior} />
           </Table.Row>
         )}
-        {getRowWithContextMenu(
+        {withContextMenu(
           <Table.Row key="3" accessibility={gridRowBehavior}>
             <Table.Cell content="3" key="1-1" accessibility={gridCellBehavior} />
             <Table.Cell content="Ali" key="1-2" accessibility={gridCellBehavior} />